added second level cache

+++ src/main/java/com/example/dogsproject/services/OwnerService.java
@@ -2,9 +2,9 @@ package com.example.dogsproject.services;
 
 import com.example.dogsproject.converter.OwnerConverter;
 import com.example.dogsproject.dto.OwnerCreateDto;
+import com.example.dogsproject.dto.OwnerResponseDto;
 import com.example.dogsproject.exceptions.AppException;
 import com.example.dogsproject.models.Owner;
-import com.example.dogsproject.repositories.owner.OwnerRepositoryHQL;
 import com.example.dogsproject.repositories.owner.OwnerRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Qualifier;
@@ -13,6 +13,7 @@ import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
+import java.util.stream.Collectors;
 
 @Service
 public class OwnerService {
@@ -20,41 +21,53 @@ public class OwnerService {
     private final OwnerConverter ownerConverter;
 
     @Autowired
-    public OwnerService(@Qualifier("ownerRepositoryHQL") OwnerRepositoryHQL ownerRepository, OwnerConverter ownerConverter) {
+    public OwnerService(@Qualifier("ownerRepositoryHQL") OwnerRepository ownerRepository, OwnerConverter ownerConverter) {
         this.ownerRepository = ownerRepository;
         this.ownerConverter = ownerConverter;
     }
 
-    public Owner save(OwnerCreateDto dto) {
+    public OwnerResponseDto save(OwnerCreateDto dto) {
         Owner owner = ownerConverter.createDtoToOwner(dto);
 
         if (owner.getDogs() != null && !owner.getDogs().isEmpty()) {
             owner.getDogs().forEach(dog -> dog.setOwner(owner));
         }
-        return ownerRepository.save(owner);
+        return ownerConverter.ownerToResponseDto(ownerRepository.save(owner));
     }
 
-    public Owner findById(Long id) {
+    public OwnerResponseDto findById(Long id) {
+        return ownerConverter.ownerToResponseDto(findOwnerById(id));
+    }
+
+    public Owner findOwnerById(Long id) {
         try {
             return ownerRepository.findById(id);
         } catch (EmptyResultDataAccessException e) {
             throw new AppException("Owner not found. Owner id: " + id, HttpStatus.NOT_FOUND);
         }
-
     }
 
-    public List<Owner> findAll() {
-        return ownerRepository.findAll();
+    public List<OwnerResponseDto> findAll() {
+        return ownerRepository
+            .findAll()
+            .stream()
+            .map(ownerConverter::ownerToResponseDto)
+            .collect(Collectors.toList());
     }
 
-    public Owner update(OwnerCreateDto dto, Long id) {
+    public OwnerResponseDto update(OwnerCreateDto dto, Long id) {
         Owner owner = ownerConverter.createDtoToOwner(dto);
         owner.setId(id);
-        return ownerRepository.save(owner);
+        return ownerConverter.ownerToResponseDto(ownerRepository.save(owner));
     }
 
-    public void delete(Owner owner) {
-        ownerRepository.delete(owner);
+    public void delete(Long ownerId) {
+        Owner owner = ownerRepository.findById(ownerId);
+        if (owner != null) {
+            ownerRepository.delete(owner);
+        } else {
+            throw new AppException("Owner not found. Owner id: " + ownerId, HttpStatus.NOT_FOUND);
+        }
     }
 
 }
